+++ src/hooks/useStateWithGetter.js
@@ -1,0 +1,19 @@
+import React, { useState, useRef } from 'react';
+
+// TODO figure out if this is bad? 
+// https://github.com/facebook/react/issues/14543
+function useStateWithGetter(initial) {
+  const ref = useRef();
+  const [state, setState] = useState(initial);
+  ref.current = state;
+  const set = (value) => {
+    ref.current = value;
+    setState(value);
+  };
+  const get = () => {
+    return ref.current;
+  };
+  return [state, set, get];
+}
+
+export default useStateWithGetter;